apps/agent_api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic==2.9.2
# Retrieval
google-cloud-discoveryengine==0.15.0
google-cloud-storage==2.18.2
# Generation
vertexai==1.70.0
# Orchestration
langchain==0.3.7
langgraph==0.2.39
